BackEnd/pyproject.toml
[project]
name = "tuteai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "uuid>=1.30",
    "uvicorn>=0.34.2",
]
